Move files into features
